Move PHON descriptors into weavy
